Eliminated Recursive Formulas from code
Eliminated both position() and comPosition() functions which were recursive and contained switch statements.
Those functions were replaced by simpler functions: fillPosition() and positionAvailable().

These two functions seperated the two tasks the original position() function had which were to check if the desired position was filled and then to place the player on that position.

Furthermore, a structure array was created which holds the position numbers (i.e. 1,2,3,4,5,6,7,8,9) and the x and y coordinates of the game board matrix.

For example, position 4 is [1][0] on the game board, so the x and y coordinates of position 4 in the structure are 1 and 0 respectively.

This is used to shorten the amount of lines of code and to assign specific combinations on the game board array to a number on the board.