Intro of Game Created, Functions and CLasses created for rest of game
The intro of the game has been programmed. Additionally, functions and classes were created for the rest of the game.

A Minerva class was created for the 4 different locations on Minerva and the special items that need to be collected to repair the ship.

A Demeter class was created for holding the status of the ship's systems that need repair.

An Options function was created to be called whenever the player needs to open the options menu.