Bring Player back to menu
When the player loses or wins the game, the player is brought back to the main menu where they can either play again, or quit the program.

This was done by change the Win_condition() function from the Ship_Inventory class from a void to a bool. This would be used in the game to determine whether or not to bring the player back to the menu.

The losing screen was done by adding a bool died() function to the Paths class that woul either return true or false. If true, the game would take the player back to the main menu. This was done by changing the deployHazard() and determineProb() functions in the Hazard class from voids to bools, which would return true or false depending on whether the player died.