Input Validation fixed
Input validation is now fixed. The program does not crash when inputting a special character (. , $ , % , * , etc.) or an alphabetical character.

This was achieved by making the input a string, then taking that string, checking to see if it had any special or alpahabetic character in the convertToInt() function. If it has anything other than a number, it prompts the user to enter a new input.

When that is resolved, it converts the string to an integer, then checks if the integer is between the accepted range.